requirements.txt
fastapi
uvicorn
python-multipart
requests
psycopg2-binary
mcp
ollama
openai-whisper
langgraph
langchain-ollama
streamlit
pandas
plotly
streamlit-autorefresh
langchain-chroma
langchain-community
pypdf
sentence-transformers
pydantic-ai